requirements.txt
# Core Dependencies
streamlit>=1.32.0
langgraph>=0.2.0
langchain-community>=0.2.0
pydantic>=2.0.0

# Google Gemini API
google-generativeai>=0.3.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Optional: Development Dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 